rename header so it does not clash with a header from binutils
