+++ src/Bfd.cpp
@@ -1,4 +1,4 @@
-#include "Bfd.h"
+#include "nmdepend.h"
 #include "ObjectFile.h"
 #include "Factory.h"
